+++ src/Exception/UnexpectedRuleException.php
@@ -19,7 +19,7 @@ use Throwable;
  *
  * final class MyRuleHandler implements RuleHandlerInterface
  * {
- *     public function validate(mixed $value, object $rule, ValidationContext $context): Result
+ *     public function validate(mixed $value, RuleInterface $rule, ValidationContext $context): Result
  *     {
  *         if (!$rule instanceof MyRule) {
  *             throw new UnexpectedRuleException(MyRule::class, $rule);