+++ src/Rule/Image/ImageHandler.php
@@ -8,6 +8,7 @@ use Psr\Http\Message\UploadedFileInterface;
 use Yiisoft\Validator\Exception\UnexpectedRuleException;
 use Yiisoft\Validator\Result;
 use Yiisoft\Validator\RuleHandlerInterface;
+use Yiisoft\Validator\RuleInterface;
 use Yiisoft\Validator\ValidationContext;
 
 use function is_string;
@@ -27,7 +28,7 @@ final class ImageHandler implements RuleHandlerInterface
         $this->imageInfoProvider = $imageInfoProvider ?? new NativeImageInfoProvider();
     }
 
-    public function validate(mixed $value, object $rule, ValidationContext $context): Result
+    public function validate(mixed $value, RuleInterface $rule, ValidationContext $context): Result
     {
         if (!$rule instanceof Image) {
             throw new UnexpectedRuleException(Image::class, $rule);