+++ src/Rule/IpHandler.php
@@ -9,6 +9,7 @@ use Yiisoft\NetworkUtilities\IpHelper;
 use Yiisoft\Validator\Exception\UnexpectedRuleException;
 use Yiisoft\Validator\Result;
 use Yiisoft\Validator\RuleHandlerInterface;
+use Yiisoft\Validator\RuleInterface;
 use Yiisoft\Validator\ValidationContext;
 
 use function is_string;
@@ -28,7 +29,7 @@ final class IpHandler implements RuleHandlerInterface
      */
     private const NEGATION_CHARACTER = '!';
 
-    public function validate(mixed $value, object $rule, ValidationContext $context): Result
+    public function validate(mixed $value, RuleInterface $rule, ValidationContext $context): Result
     {
         if (!$rule instanceof Ip) {
             throw new UnexpectedRuleException(Ip::class, $rule);