+++ src/Rule/StringValueHandler.php
@@ -7,6 +7,7 @@ namespace Yiisoft\Validator\Rule;
 use Yiisoft\Validator\Exception\UnexpectedRuleException;
 use Yiisoft\Validator\Result;
 use Yiisoft\Validator\RuleHandlerInterface;
+use Yiisoft\Validator\RuleInterface;
 use Yiisoft\Validator\ValidationContext;
 
 use function is_string;
@@ -18,7 +19,7 @@ use function is_string;
  */
 final class StringValueHandler implements RuleHandlerInterface
 {
-    public function validate($value, object $rule, ValidationContext $context): Result
+    public function validate($value, RuleInterface $rule, ValidationContext $context): Result
     {
         if (!$rule instanceof StringValue) {
             throw new UnexpectedRuleException(StringValue::class, $rule);