+++ tests/Support/Rule/StubRule/StubRuleHandler.php
@@ -6,11 +6,12 @@ namespace Yiisoft\Validator\Tests\Support\Rule\StubRule;
 
 use Yiisoft\Validator\Result;
 use Yiisoft\Validator\RuleHandlerInterface;
+use Yiisoft\Validator\RuleInterface;
 use Yiisoft\Validator\ValidationContext;
 
 final class StubRuleHandler implements RuleHandlerInterface
 {
-    public function validate(mixed $value, object $rule, ValidationContext $context): Result
+    public function validate(mixed $value, RuleInterface $rule, ValidationContext $context): Result
     {
         return new Result();
     }